api/requirements.txt
fastapi==0.135.1
uvicorn==0.34.2
langchain==1.2.12
langchain-anthropic==1.3.5
langgraph==1.1.2
langsmith==0.7.17
python-dotenv==1.1.0
pydantic==2.11.3
